Disable rotational smoothing and instead reset physics on large rotations.
In vanilla Skyrim the player character doesn't turn smoothly if you use mouse+KB, rotate the camera, and push forward to turn, you just turn instantly. This is easiest to see if you rotate the camera 180 and turn completely around. This creates a large amount of movement which could create undesired effects in SMP. hydrogen's fix was to clamp the skeleton's rotation, which causes the character model rotate over time instead of instantly turning. Unfortunately it appears that attached weapon nodes aren't captured by the rotationial clamping and move instantly still, creating a bad looking effect. This commit disables rotational clamping and instead resets the actor that rotated's physics on the rotation frame, so the player will turn instantly again, and it won't cause major movement in the physics world.