Don't overwrite armors that already exist with valid physics
This occurs when the same prefix is generated for two armors, often
the body and one of the KSSMP hairs. Ideally we'd fix the prefix
generation code, but it's good to handle the collision situation
anyway.

Note that this bug existed in older versions as well, but they were
less strict about removing armors so the older ones just remained
permanently attached to the Bullet world. They continued to work,
so the bug wasn't visible, except for a gradual loss of FPS.