All-new mesh remapping feature in defaultBBPs.xml
This makes it possible to give consistent names to meshes at the level
of the defaults file, so only one collision body definition is needed no
matter how inconsistent the original models are. The priority option
(similar to disable-priority in collision body definitions) gives extra
control over which meshes are renamed.

This also supports multiple meshes per collision body, so it's possible
to use one collision body definition for armors that are made up of
several differently-named meshes.