Refactored collision check struct for better flexibility
Added new collision algorithm, which builds lists for A and B in advance
and updates bounding boxes in the process. This does more work in the
linear phase, but usually ends up with smaller lists in the quadratic
phase. Looks like it may give a performance boost in CPU, but its real
purpose is to set things up better for GPU collision checking.