Do the bounding box reduction on the GPU.
This means we can get the aggregate bounding boxes early, and start some
further processing on the CPU before the individual collider bounding
boxes are ready. Eventually we should be able to avoid transferring the
individual boxes altogether.

Performance still degrades significantly when there are a lot of NPCs
around - suspect this is to do with the CUDA queue depth rather than raw
bandwidth.