Do some of the collision checks at compile time
Sphere-triangle collision has a number of cases depending on the sign of
the penetration parameter. Since this is always the same across a shape,
we can have separate kernels for the three cases and select before
launch.

This only applies to the CUDA implementation - the CPU one still does
all the checks at runtime per collision pair.