Do part of the bounding box check on GPU
This does the bounding box check between shape A colliders with the
shape B bounding box. It currently uses scratch space defined on shape
A, which works but isn't great for performance as that space is limited.
Moving it to shared memory might be better.

Also changed the reduce algorithm to one that uses warp-level primitives
and a fraction of the amount of shared memory. The reduce intrinsics
would be even better, but require a 30 series GPU.