Use shared memory for vertex lists instead of scratch space
Vertex lists are usually small and should fit in shared memory without
any trouble. We allow up to 256 vertices (or triangles) in each list -
they're typically around 20-50. This also removes the need to run the
collision kernel in multiple batches.

The collision kernel can now compute lists for both objects. It will do
the larger one first, and decide whether to create a list for the
smaller one or just brute-force all the collisions.

Performance is about the same as it was with lists calculated on the
CPU (but far better than when it was using scratch space in global
memory).