Reworked collision scheduling for CUDA
Each collision is now processed in two parts, so we can launch all the
kernels first before having to synchronize anything (note that both
parts are still fairly CPU-heavy, so they're parallelized as well).

This mechanism is only used for GPU - CPU collision still uses the
previous mechanism, as it doesn't have any blocking functions that would
benefit from this kind of split.