New version of planar struct templates
This version uses fewer temporaries and less recursion when assigning
from a real struct, in the hope of making it easier for the compiler to
optimize. It also adds support for assigning directly between planar
arrays without an intermediate struct.

In practice, it seems to make absolutely no difference at all to
register use, code size, or performance. So the compiler was doing just
fine already. Oh well.