Reworked planar pointer storage
We now store a separate pointer for each plane, instead of computing it
from the base every time. This gets the expected reduction in register
usage on the GPU (should be the same as with interleaved structs).

Per-vertex and per-triangle updates now use planar arrays for both
inputs and outputs (plus the inevitable random access to vertex data).