Added template for interleaved struct arrays
This is basically just a regular array, but with an interface compatible
with the planar one. Added ArrayType definition so it's easy to switch
between the two. Aside from register use and code size, there's very
little real difference (if any kernel is going to benefit from planar
access, it's probably the body update, which hasn't been converted yet).

Removed the "external" penetration type, since it's equivalent to the
internal case with any two vertices reversed. This can be done at
setup time and eliminates one of the CUDA kernels.