Moved internal update logic to a GPU kernel
We now use dynamic parallelism to launch the 3 or 5 kernels required for
internal updates. This removes most of the concerns about order of
kernel launches from the dispatcher. It also means vertex and triangle
updates can be done concurrently on separate streams.

On 6 cores, internal updates are now similar or better performance on
GPU than CPU.