More sophisticated memory layout for merge buffer
The merge buffer now only contains entries for bone pairs where at least
one is dynamic.

This significantly reduces transfer time when doing collisions between
bodies with a lot of kinematic bones.

Theoretically it can increase the merge buffer size if most bones of
both bodies are dynamic (because those pairs are duplicated), but that's
unlikely to happen with most setups.