Don't run multiple iterations for slow frames
If one frame takes a very long time (for example, because it's setting
up new NPCs), this could cause the engine to run up to 5 iterations of
physics on the next frame. If collision and physics processing time is
non-negligible (more than about 8ms), this causes a feedback loop and
reduces framerate for an extended time.

This should no longer happen. The code for deciding how many iterations
to run and how long they should be needs some proper rework, but this
will do for now. It can still run multiple iterations if the framerate
is very high, but I'm not sure how often that actually happens.