Dynamic substep timing
Simple proportional control. This no longer assumes physics calculations
are (almost) free. It should react fast to slow frames, dropping the
number of iterations to 1, then gradually return to shorter intervals
and more iterations.

If min-fps is significantly higher than the framerate cap, this will
effectively disable control and always do 1 iteration. If min-fps is
lower than the framerate cap, it will increase iterations to obtain that
framerate.

Note that if min-fps is MUCH lower than the framerate cap, physics
timing is likely to get messed up. So don't do that.