Delayed application of collision results
Triangle-vertex and vertex-vertex collision results are now applied on
the next frame instead of the current one. This reduces accuracy, but
allows GPU collision to run concurrent with the rest of the game engine,
instead of leaving the CPU idle.

Triangle-triangle collisions are assumed to need full accuracy, and are
still applied on the current frame.