Further work on simplifying kernel parameters
Note that none of this actually reduces the amount of data transferred
(although that may now be possible), just simplifies the functions calls
for my sanity. It does seem to reduce binary size.

We no longer need to replicate margin and penetration values for every
collider, and instead have them per-shape, the same way they're used in
the CPU code.