When fps is under min-fps, removed the physics simulations where the timestep is different/lower from/than min-fps.
This results in more stability when the fps are under min-fps as the
physics run always exactly at min-fps frequency, and the bullet engine
requires a stable frequency to avoid artefacts like resonance.

And results in a better performance, as no time is spent on simulating
smaller increment of time.

Consequently, a slowdown of the physics simulation compared to the game
clock exists at the frame scale, but this is corrected the next frame,
and is unnoticeable to the player; more unnoticeable than the physics
errors introduced by irregular timesteps.