Refactored the skeletons sorting algorithm preceding the choice of which skeletons are active, and what follows:
- We now sort the skeletons depending on the angle and distance.

If one is exacly on the side of the camera (product of cos = 0)
then the one equal to zero is last.
If one is behind the camera and the other in front of the camera (product of cos < 0)
then the one behind the camera is last (the one with cos(angle) < 0).
If both are on the same side of the camera (product of cos > 0)
We want first the smallest angle (so the highest cosinus), and the smallest distance,
so we want the smallest distance / cosinus.
cl = cosinus * distance, dl = distance² => distance / cosinus = dl/cl
So we want dl/cl < dr/cr.
return (cl * cr == 0) ? (cl != 0) : (cl * cr < 0) ? (cl > cr) : (dl * cr < dr * cl);

- Disabled the second call of isPlayerCharacter() during isInPlayerView().

- We don't enable the skeletons behind the camera *or on its side*.