Managing the case of the 1st person player.
We now sort the skeletons depending on the angle and distance.

If one is at distance zero (1st person player) from the camera
then it is first.
If one is exacly on the side of the camera (product of cos = 0)
then the one equal to zero is last.
If one is behind the camera and the other in front of the camera (product of cos < 0)
then the one behind the camera is last (the one with cos(angle) < 0).
If both are on the same side of the camera (product of cos > 0)
We want first the smallest angle (so the highest cosinus), and the smallest distance,
so we want the smallest distance / cosinus.
cl = cosinus * distance, dl = distance² => distance / cosinus = dl/cl
So we want dl/cl < dr/cr.
return (dl * dr == 0) ? (dl == 0) : (cl * cr == 0) ? (cl != 0) : (cl * cr < 0) ? (cl > cr) : (dl * cr < dr * cl);